[Bronze V] Title: Multiply, Time: 180 ms, Memory: 8548 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.cs
@@ -1,7 +1,9 @@
+using System.Numerics;
+
 using var sr = new StreamReader(new BufferedStream(Console.OpenStandardInput()));
 using var sw = new StreamWriter(new BufferedStream(Console.OpenStandardOutput()));
 
 sr.ReadLine();
-ulong a = ulong.Parse(sr.ReadLine());
-ulong b = ulong.Parse(sr.ReadLine());
+BigInteger a = BigInteger.Parse(sr.ReadLine());
+BigInteger b = BigInteger.Parse(sr.ReadLine());
 sw.WriteLine(a*b);
